+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/BaseServiceWriter.kt
@@ -32,7 +32,7 @@ internal abstract class BaseServiceWriter<TService : BaseService>(
      * Write the service specific objects to the database.
      *
      * @receiver The [BaseService] to write to the database.
-     * @return true if the objects were successfully writen to the database, otherwise false
+     * @return true if the objects were successfully written to the database, otherwise false
      */
     protected abstract fun TService.writeService(): Boolean
 
@@ -42,9 +42,9 @@ internal abstract class BaseServiceWriter<TService : BaseService>(
      * @receiver The [BaseService] to write to the database.
      * @param T The type of object to write to the database.
      * @param writer The callback used to write the objects to the database. Will be called once for each object and should return true if the object is
-     *   successfully writen to the database.
+     *   successfully written to the database.
      *
-     * @return true if all objects are successfully writen to the database, otherwise false.
+     * @return true if all objects are successfully written to the database, otherwise false.
      */
     protected inline fun <reified T : IdentifiedObject> TService.writeEach(noinline writer: (T) -> Boolean): Boolean {
         var status = true
@@ -53,13 +53,13 @@ internal abstract class BaseServiceWriter<TService : BaseService>(
     }
 
     /**
-     * Validate that an object is actually writen to the database, logging an error if anything goes wrong.
+     * Validate that an object is actually written to the database, logging an error if anything goes wrong.
      *
-     * @param T The type of object being writen.
-     * @param it The object being writen.
+     * @param T The type of object being written.
+     * @param it The object being written.
      * @param writer The callback actually saving the object to the database.
      *
-     * @return true if the object is successfully writen to the database, otherwise false.
+     * @return true if the object is successfully written to the database, otherwise false.
      */
     protected inline fun <reified T : IdentifiedObject> validateWrite(it: T, noinline writer: (T) -> Boolean): Boolean {
         return validateWrite(it, writer) { e ->