+++ src/main/kotlin/com/zepben/evolve/database/sqlite/extensions/ConnectionExtensions.kt
@@ -10,7 +10,7 @@ package com.zepben.evolve.database.sqlite.extensions
 
 import java.sql.Connection
 
-fun Connection.configureBatch(): Connection {
+internal fun Connection.configureBatch(): Connection {
     createStatement().use { statement ->
         statement.executeUpdate("PRAGMA journal_mode = OFF")
         statement.executeUpdate("PRAGMA synchronous = OFF")