+++ src/test/kotlin/com/zepben/evolve/services/common/translator/TranslatorTestBase.kt
@@ -71,7 +71,7 @@ internal abstract class TranslatorTestBase<S : BaseService>(
             val actual = validationInfo.map { it.cim::class.simpleName!! }.toSet()
             val expected = (databaseTables.tables.keys - excludedTables).map { it.simpleName!!.removePrefix("Table").removeSuffix("s") }.toSet()
             fail(
-                "The number of items being validated did not match the number of items writen to the database. Did you forget to validate an item, " +
+                "The number of items being validated did not match the number of items written to the database. Did you forget to validate an item, " +
                     "or to exclude the table if it was an association or array data?" +
                     formatValidationError("Unexpected", actual - expected) +
                     formatValidationError("Missing", expected - actual)