+++ changelog.md
@@ -27,14 +27,15 @@
   * `load` has been renamed to `read`, so the readers now read, rather than load.
 * Database readers and writers no longer have the container of the data they will read/write passed to the constructor. They now have this passed to the `read`
   or `write` method.
-* The following classes are now internal:
+* The following classes and methods are now internal:
   * `BaseServiceReader`, `BaseServiceWriter`, `BaseCollectionReader`, `BaseCollectionWriter`, `BaseEntryWriter`, 
   * `CimReader`, `CimWriter`,
   * `CustomerCimReader`, `CustomerCimWriter`, `CustomerServiceReader`, `CustomerServiceWriter`, 
   * `DiagramCimReader`, `DiagramCimWriter`, `DiagramServiceReader`, `DiagramServiceWriter`, 
   * `MetadataCollectionReader`, `MetadataCollectionWriter`, `MetadataEntryReader`, `MetadataEntryWriter`, 
-  * `MetricsEntryWriter`, `MetricsWriter`
+  * `MetricsEntryWriter`, `MetricsWriter`,
   * `NetworkServiceReader`, `NetworkServiceWriter`,
+  * Extension methods in `com.zepben.evolve.database.sqlite.extensions` (e.g. `ResultSet.getNullableDouble`)
 * The following classes now have internal constructors:
   * `BaseDatabaseWriter`, `CimDatabaseReader`, `CimDatabaseWriter`,
   * `CimDatabaseTables`, `BaseDatabaseTables`, `CustomerDatabaseTables`, `DiagramDatabaseTables`, `NetworkDatabaseTables`, `MetricsDatabaseTables`